switched to solita, changed on-chain representation to i64 for the decimal numbers

+++ program/src/stats_bucket_instruction.rs
@@ -13,7 +13,7 @@ pub struct CombineParams {
 #[derive(BorshDeserialize, BorshSerialize, Debug, Clone)]
 pub struct PushParams {
     pub name : String,
-    pub ys : Vec<f32>
+    pub ys : Vec<i64>
 }
